Move settings to app site
