+++ src/Content/App/App.tsx
@@ -1,4 +1,4 @@
-import { Child } from 'src/Global/Decorators';
+import { Child, PageScriptListener } from 'src/Global/Decorators';
 import { settings } from 'src/Content/Runtime/Settings';
 
 @Child
@@ -14,11 +14,19 @@ export class App {
 		const map = [
 			['7tv.webp', chrome.runtime.getURL('image/7tv.webp')],
 			['7tv-nd.webp', chrome.runtime.getURL('image/7tv-nd.webp')],
+			['twitter.webp', chrome.runtime.getURL('image/icon/twitter.webp')],
+			['discord.webp', chrome.runtime.getURL('image/icon/discord.webp')],
 			['stylesheet', chrome.runtime.getURL('styles/Style.css')]
 		] as [string, string][];
 
 		this.sendMessageDown('ConfigChange', settings.raw);
 		this.sendMessageDown('OnAssets', map);
+	}
+
+	@PageScriptListener('OpenConfigMenu')
+	whenOpenTheConfigMenu(): void {
+		const overlay = document.querySelector('.seventv-overlay');
+		console.log('open overlay', overlay);
 	}
 
 	/**
